+++ api/src/main/java/com/lifegraph/team20/payload/response/ChildGraphReference.java
@@ -3,7 +3,7 @@ package com.lifegraph.team20.payload.response;
 import lombok.Data;
 
 @Data
-public class ChildGraph {
+public class ChildGraphReference {
 
   private Long id;
   private Long parent_id;
@@ -11,7 +11,7 @@ public class ChildGraph {
   private int score;
   private String comment;
 
-  public ChildGraph(Long id, Long parent_id, int age, int score, String comment) {
+  public ChildGraphReference(Long id, Long parent_id, int age, int score, String comment) {
 
     this.id = id;
     this.parent_id = parent_id;
