+++ api/src/main/java/com/lifegraph/team20/security/WebSecurityConfig.java
@@ -57,9 +57,9 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 		http.cors().and().csrf().disable()
 			.exceptionHandling().authenticationEntryPoint(unauthorizedHandler).and()
 			.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
-			.authorizeRequests().antMatchers("/auth/**", "/test/**", "/accountReference/**", "/reference/**", "/search/**","/api/**").permitAll()
-			.antMatchers("/life-graphs/**").permitAll()
-			.anyRequest().authenticated();
+			.authorizeRequests().antMatchers("/auth/**").permitAll()
+			.antMatchers("/test/**").permitAll()
+		            .anyRequest().authenticated();
 
     http.addFilterBefore(authenticationJwtTokenFilter(), UsernamePasswordAuthenticationFilter.class);
   }
